Did some more changes
Replaced the emplace function with a working insert.  Wasn't sure if
since insert is on our slides if she really wanted us to work with that
function.  So i fixed it and got it to work anyway.

Went a bit O.C.D fixing spacing in all the files

made slight change to some if else statements that were using the end()
flag within a for loop also using the end() flag.  (we should ask if
end()-1 is acceptable since end() is a flag after all...)

went into input/outputhandler.h files and edited the parameters to
changed them from StudentList to studentListChild class objects (you
guys included the header but must of missed the params in those so got
them fixed.)

I also removed all the old commented code we originally left in.  So now
everything looks crisp and clean.